+++ src/python/grpcio_tests/tests_aio/unit/aio_rpc_error_test.py
@@ -19,7 +19,7 @@ import unittest
 import grpc
 
 from grpc.experimental import aio
-from grpc.experimental.aio._call import AioRpcError
+from grpc.aio._call import AioRpcError
 from tests_aio.unit._test_base import AioTestBase
 
 _TEST_INITIAL_METADATA = aio.Metadata(
